moving code around to prep for changes
