+++ pkg/runner/client.go
@@ -1,6 +1,7 @@
 // This file and its contents are licensed under the Apache License 2.0.
 // Please see the included NOTICE for copyright information and
 // LICENSE for a copy of the license.
+
 package runner
 
 import (
@@ -23,8 +24,7 @@ import (
 )
 
 var (
-	appVersion         = pgmodel.VersionInfo{Version: version.Promscale, CommitHash: version.CommitHash}
-	migrationLockError = fmt.Errorf("Could not acquire migration lock. Ensure there are no other connectors running and try again.")
+	appVersion = pgmodel.VersionInfo{Version: version.Promscale, CommitHash: version.CommitHash}
 )
 
 func CreateClient(cfg *Config, promMetrics *api.Metrics) (*pgclient.Client, error) {
@@ -73,9 +73,9 @@ func CreateClient(cfg *Config, promMetrics *api.Metrics) (*pgclient.Client, erro
 		if !cfg.UseVersionLease {
 			lease = nil
 		}
-		err = SetupDBState(conn, appVersion, lease, extOptions)
-		migrationFailedDueToLockError = err == migrationLockError
-		if err != nil && err != migrationLockError {
+		err = pgmodel.Migrate(conn, appVersion, lease, extOptions)
+		migrationFailedDueToLockError = err == pgmodel.MigrationLockError
+		if err != nil && err != pgmodel.MigrationLockError {
 			return nil, fmt.Errorf("migration error: %w", err)
 		}
 
@@ -224,44 +224,6 @@ func isBGWLessThanDBs(conn *pgx.Conn) (bool, error) {
 		return true, nil
 	}
 	return false, nil
-}
-
-func SetupDBState(conn *pgx.Conn, appVersion pgmodel.VersionInfo, leaseLock *util.PgAdvisoryLock, extOptions extension.ExtensionMigrateOptions) error {
-	// At startup migrators attempt to grab the schema-version lock. If this
-	// fails that means some other connector is running. All is not lost: some
-	// other connector may have migrated the DB to the correct version. We warn,
-	// then start the connector as normal. If we are on the wrong version, the
-	// normal version-check code will prevent us from running.
-
-	if leaseLock != nil {
-		locked, err := leaseLock.GetAdvisoryLock()
-		if err != nil {
-			return fmt.Errorf("error while acquiring migration lock %w", err)
-		}
-		if !locked {
-			return migrationLockError
-		}
-		defer func() {
-			_, err := leaseLock.Unlock()
-			if err != nil {
-				log.Error("msg", "error while releasing migration lock", "err", err)
-			}
-		}()
-	} else {
-		log.Warn("msg", "skipping migration lock")
-	}
-
-	err := pgmodel.Migrate(conn, appVersion, extOptions)
-	if err != nil {
-		return fmt.Errorf("Error while trying to migrate DB: %w", err)
-	}
-
-	_, err = extension.InstallUpgradePromscaleExtensions(conn, extOptions)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func ApplyDatasetConfig(conn *pgx.Conn, cfgFilename string) error {
